add `--dry-run` option

+++ cli/driver/src/exporter.rs
@@ -320,34 +320,38 @@ impl Callbacks for ExtractionCallbacks {
                             ),
                         );
                     }
-                    let metadata = cargo_metadata::MetadataCommand::new().exec().unwrap();
-                    let crate_name = std::env::var("CARGO_CRATE_NAME").unwrap();
-                    let package = metadata
-                        .packages
-                        .iter()
-                        .rfind(|pkg| pkg.name == crate_name)
-                        .unwrap();
-                    let manifest_path = package.manifest_path.clone();
-                    let relative_path: cargo_metadata::camino::Utf8PathBuf =
-                        ["hax", format!("{}", backend.backend).as_str(), "extraction"]
+                    if backend.dry_run {
+                        serde_json::to_writer(std::io::stdout(), &output.files).unwrap();
+                    } else {
+                        let metadata = cargo_metadata::MetadataCommand::new().exec().unwrap();
+                        let crate_name = std::env::var("CARGO_CRATE_NAME").unwrap();
+                        let package = metadata
+                            .packages
                             .iter()
-                            .collect();
-                    let output_dir = manifest_path.parent().unwrap().join(relative_path);
-                    for file in output.files.clone() {
-                        let path = output_dir.join(file.path);
-                        std::fs::create_dir_all({
-                            let mut parent = path.clone();
-                            parent.pop();
-                            parent
-                        })
-                        .unwrap();
-                        session.note_without_error(format!("Writing file {:#?}", path));
-                        std::fs::write(&path, file.contents).unwrap_or_else(|e| {
-                            session.fatal(format!(
-                                "Unable to write to file {:#?}. Error: {:#?}",
-                                path, e
-                            ))
-                        })
+                            .rfind(|pkg| pkg.name == crate_name)
+                            .unwrap();
+                        let manifest_path = package.manifest_path.clone();
+                        let relative_path: cargo_metadata::camino::Utf8PathBuf =
+                            ["hax", format!("{}", backend.backend).as_str(), "extraction"]
+                                .iter()
+                                .collect();
+                        let output_dir = manifest_path.parent().unwrap().join(relative_path);
+                        for file in output.files.clone() {
+                            let path = output_dir.join(file.path);
+                            std::fs::create_dir_all({
+                                let mut parent = path.clone();
+                                parent.pop();
+                                parent
+                            })
+                            .unwrap();
+                            session.note_without_error(format!("Writing file {:#?}", path));
+                            std::fs::write(&path, file.contents).unwrap_or_else(|e| {
+                                session.fatal(format!(
+                                    "Unable to write to file {:#?}. Error: {:#?}",
+                                    path, e
+                                ))
+                            })
+                        }
                     }
                 }
             };
